http2: improve error when server sends HTTP/1
If for some reason the server sends an HTTP/1.1 response response
starting with "HTTP/1.1 " (9 bytes), http2 transport interprets that as
a valid frame header for an expected payload of ~4MB, and fails with
non-descriptive messages like "unexpected EOF".

This could happen, for example, if ALPN is misconfigured on the server's
load balancer.

This change attempts to improve that feedback by noting in the error
messages whenever the frame header was exactly the "HTTP/1.1 bytes".

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>